backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-storage==2.10.0
google-cloud-firestore==2.13.0
httpx==0.25.0
python-multipart==0.0.6
pydantic==2.5.0
